% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeBill.R
\name{BillIDfromURL}
\alias{BillIDfromURL}
\title{Get Bill ID from URL}
\usage{
BillIDfromURL(url)
}
\arguments{
\item{url}{\code{character} scalar of vector of Bill Summary Page URL(s)}
}
\value{
\code{character} scalar of Bill ID.
}
\description{
Get Bill ID from URL
}
