% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectBillData.R
\name{collectBillData}
\alias{collectBillData}
\title{Collecting Bill Information from a \code{scrapeBill} Object.}
\usage{
collectBillData(x, include = c("BillID", "URL", "Title", "Summary",
  "Sponsorship", "Action", "Committee", "Subject"), expand = "None",
  progress = FALSE)
}
\arguments{
\item{x}{A (or A list of) \code{scrapeBill} object created by \code{\link{scrapeBill}} function.}

\item{include}{\code{character} vector of What element(s) of bill information to 
include in output. Options are \code{"BillID"}, \code{"URL"}, \code{"Title"},
\code{"Summary"}, \code{"Sponsorship"}, \code{"Action"}, \code{"Committee"},
\code{"Subject"}, and \code{"Text"}. The default is to include everything 
except for \code{"Text"}. Need at least one element. 
Information set by this option uses single row output 
(i.e., Set \code{officialonly}, \code{latestonly}, \code{single}, and 
\code{allinone} to \code{TRUE} in \code{collectOPTION} functions).}

\item{expand}{The default is \code{"None"} (not to expand data). Else,
create long data by using expanded information (i.e., Set \code{officialonly}, \code{latestonly}, \code{single}, and 
\code{allinone} to \code{FALSE} in \code{collectOPTION} functions). 
Choose option from ONE of \code{"Title"},
\code{"Summary"}, \code{"Sponsorship"}, \code{"Action"}, \code{"Committee"},
\code{"Subject"}, and \code{"Text"}. CANNOT set more than one option.}

\item{progress}{If \code{TRUE}, show progress bar for the processing of multiple bills.}
}
\value{
A \code{data.frame} object of bill information.
}
\description{
Collect Bill Information and Put Into Data Frame.
}
\examples{
tgturl <- "https://www.congress.gov/bill/116th-congress/senate-bill/252"
bill252 <- scrapeBill(tgturl)
collectBillData(bill252)

}
