% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectBillData.R
\name{collectBillID}
\alias{collectBillID}
\title{Collecting Bill ID from a \code{scrapeBill} Object.}
\usage{
collectBillID(x, progress = FALSE)
}
\arguments{
\item{x}{A (or A list of) \code{scrapeBill} object created by \code{\link{scrapeBill}} function.}

\item{progress}{If \code{TRUE}, show progress bar for the processing of multiple bills.}
}
\value{
A \code{data.frame} with bill ID information.
}
\description{
Collect information that identifies bill.
}
\examples{
tgturl <- "https://www.congress.gov/bill/109th-congress/house-bill/247"
bill247 <- scrapeBill(tgturl)
collectBillID(bill247)

}
