% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectBillData.R
\name{collectCommittee}
\alias{collectCommittee}
\title{Collecting Committee Information from a \code{scrapeBill} Object.}
\usage{
collectCommittee(x, single = TRUE, progress = FALSE)
}
\arguments{
\item{x}{A (or A list of) \code{scrapeBill} object created by \code{\link{scrapeBill}} function.}

\item{single}{Whether to export single row results (see following description). 
The default is \code{TRUE}.}

\item{progress}{If \code{TRUE}, show progress bar for the processing of multiple bills.}
}
\value{
If \code{single==TRUE}, returns \code{data.frame} object with 
following information (single row):
\itemize{
  \item \code{Committee_N}: Number of Committees the Bill Referred to
  \item \code{Committee_Names}: Name(s) of Referred Committees (If multiple committees, divided by \code{;}) 
}

If \code{single==FALSE}, A \code{data.frame} object containing following variables 
(rows represent date of activity):
\itemize{
  \item \code{Date}: Date of Committee Activity
  \item \code{Main}: Name of Committee 
  \item \code{Sub}: Name of Subcommittee, if any
  \item \code{Activity}: Content of Activity 
  \item \code{Reports}: Name of Reports, if any 
}
}
\description{
Collect information regarding committees considering the bill.
}
\examples{
tgturl <- "https://www.congress.gov/bill/116th-congress/senate-bill/252"
bill252 <- scrapeBill(tgturl)
collectCommittee(bill252)

}
