% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectBillData.R
\name{collectSponsorship}
\alias{collectSponsorship}
\title{Collecting Sponsorship Information from a \code{scrapeBill} Object.}
\usage{
collectSponsorship(x, single = TRUE, progress = FALSE)
}
\arguments{
\item{x}{A (or A list of) \code{scrapeBill} object created by \code{\link{scrapeBill}} function.}

\item{single}{Whether to export single row results (see following description). 
The default is \code{TRUE}.}

\item{progress}{If \code{TRUE}, show progress bar for the processing of multiple bills.}
}
\value{
If \code{single==TRUE}, returns \code{data.frame} object with 
following variables (single row):
\itemize{
  \item \code{Sp_Chamber}: Chamber of Primary Sponsor
  \item \code{Sp_Name}: Full Name of Primary Sponsor
  \item \code{Sp_LastName}: Last Name of Primary Sponsor
  \item \code{Sp_Party}: Party of Primary Sponsor
  \item \code{Sp_State}: State of Primary Sponsor
  \item \code{Sp_District}: District of Primary Sponsor (Only for House Representatives)
  \item \code{Cosp_N}: Total Number of Cosponsors
  \item \code{Cosp_Original}: Number of Original Cosponsors
  \item \code{Cosp_D}: Number of Democratic Cosponsors
  \item \code{Cosp_R}: Number of Republican Cosponsors
}

If \code{single==FALSE}, \code{data.frame} object with following 
variables (rows represent sponsors):
\itemize{
  \item \code{Date}: Date of Sponsorship. If Primary Sponsor, this is the bill introduction date.
  \item \code{Status}: Status of Sponsorship. One of "Primary Sponsor," "Original Cosponsor," and "Cosponsor."
  \item \code{Chamber}: Chamber of Sponsor
  \item \code{Name}: Full Name of Sponsor
  \item \code{LastName}: Last Name of Sponsor
  \item \code{Party}: Party of Sponsor
  \item \code{State}: State of Sponsor
  \item \code{District}: District of Sponsor (Only for House Representatives)
}
}
\description{
Collect information regarding sponsorship of the bill.
}
\examples{
tgturl <- "https://www.congress.gov/bill/109th-congress/house-bill/19"
bill19 <- scrapeBill(tgturl)
collectSponsorship(bill19)

}
\seealso{
\code{\link{scrapeBill}} and \code{\link{getscrapeBill}}
}
