% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectBillData.R
\name{collectSubject}
\alias{collectSubject}
\title{Collecting Subject Information from a \code{scrapeBill} Object.}
\usage{
collectSubject(x, single = TRUE, progress = FALSE)
}
\arguments{
\item{x}{A (or A list of) \code{scrapeBill} object created by \code{\link{scrapeBill}} function.}

\item{single}{If \code{TRUE} (default), the function returns single row output (see below).
If \code{FALSE}, long format \code{data.frame} object with separate subject fields as rows.}

\item{progress}{If \code{TRUE}, show progress bar for the processing of multiple bills.}
}
\value{
If \code{single==TRUE}, returns \code{data.frame} object with 
following information (single row):
\itemize{
  \item \code{Subject_Primary}: Primary Subject Field of the Bill
  \item \code{SUbject_N}: Number of Subject Fields Assigned to the Bill
  \item \code{Subject_All}: All Subject Fields divided by ";"
}

If \code{single==FALSE}, A \code{data.frame} object containing following variables 
(rows represent each subject filed):
\itemize{
  \item \code{ID}: Bill ID 
  \item \code{Subject}: Name of Subject Field
  \item \code{Rank}: The order of subject fields as presented in website. 1 indicates primary subject.
}
}
\description{
Collect information regarding subject field of the bill.
}
\examples{
tgturl <- "https://www.congress.gov/bill/116th-congress/senate-bill/252"
bill252 <- scrapeBill(tgturl)
collectSubject(bill252)

}
