% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectBillData.R
\name{collectSummary}
\alias{collectSummary}
\title{Collecting Summary Information from a \code{scrapeBill} Object.}
\usage{
collectSummary(x, latestonly = TRUE, progress = FALSE)
}
\arguments{
\item{x}{A (or A list of) \code{scrapeBill} object created by \code{\link{scrapeBill}} function.}

\item{latestonly}{If \code{TRUE} (default), the function returns only the text of 
the latest summary. 
If \code{FALSE}, the function returns the \code{data.frame} object of all 
published summaries and their published dates and timings.}

\item{progress}{If \code{TRUE}, show progress bar for the processing of multiple bills.}
}
\value{
A \code{character} or \code{data.frame} object with summary texts.
}
\description{
Collect information regarding summary of the bill.
}
\examples{
tgturl <- "https://www.congress.gov/bill/109th-congress/house-bill/247"
bill247 <- scrapeBill(tgturl)
collectSummary(bill247)

}
\seealso{
\code{\link{scrapeBill}} and \code{\link{getscrapeBill}}
}
