% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectBillData.R
\name{collectText}
\alias{collectText}
\title{Collecting Text Information from a \code{scrapeBill} Object.}
\usage{
collectText(x, allinone = TRUE, progress = FALSE)
}
\arguments{
\item{x}{A (or A list of) \code{scrapeBill} object created by \code{\link{scrapeBill}} function.}

\item{allinone}{If \code{TRUE} (default), the function returns all texts of 
bill into one character scalar. If \code{FALSE}, the function separates text by 
potential sections and exported as list.}

\item{progress}{If \code{TRUE}, show progress bar for the processing of multiple bills.}
}
\value{
A \code{character} or \code{list} object of Bill Texts.
}
\description{
Collect information regarding Latest Full Text of the bill.
}
\examples{
tgturl <- "https://www.congress.gov/bill/116th-congress/senate-bill/252"
bill252 <- scrapeBill(tgturl)
collectText(bill252)

}
