% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectBillData.R
\name{collectTitle}
\alias{collectTitle}
\title{Collecting Bill Title(s) from a \code{scrapeBill} Object.}
\usage{
collectTitle(x, officialonly = TRUE, progress = FALSE)
}
\arguments{
\item{x}{A (or A list of) \code{scrapeBill} object created by \code{\link{scrapeBill}} function.}

\item{officialonly}{If \code{TRUE} (default), the function returns the the text of the official title as introduced. 
If \code{FALSE}, the function returns the \code{data.frame} object of all 
published official and short titles and their published timings.}

\item{progress}{If \code{TRUE}, show progress bar for the processing of multiple bills.}
}
\value{
A \code{character} vector or \code{data.frame} list with bill title information.
}
\description{
Collect information regarding title(s) of the bill.
}
\examples{
tgturl <- "https://www.congress.gov/bill/109th-congress/house-bill/247"
bill247 <- scrapeBill(tgturl)
collectTitle(bill247)

}
\seealso{
\code{\link{scrapeBill}} and \code{\link{getscrapeBill}}
}
