% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeBill.R
\name{scrapeBill}
\alias{scrapeBill}
\title{Downloading Bill HTML files from Congress Website}
\usage{
scrapeBill(url, progress = TRUE)
}
\arguments{
\item{url}{A \code{character} object of Bill Summary Page URL (e.g., \code{https://www.congress.gov/bill/109th-congress/house-bill/247})}

\item{progress}{If \code{TRUE}(default), show progress bar for the processing of multiple bills.}
}
\value{
A (or a list of) \code{scrapeBill} object(s) that contains two \code{xml_document} objects.
\itemize{
  \item \code{info}: contains meta information of the bill
  \item \code{text}: contains full texts of the bill
}
}
\description{
Downloading Bill Information from \href{https://www.congress.gov}{US Congress Website}.
}
\examples{
tgturl <- "https://www.congress.gov/bill/109th-congress/house-bill/247"
bill247 <- scrapeBill(tgturl)

}
\seealso{
\code{\link{getBill}}, \code{\link{read_scrapeBill}}, and \code{\link{write_scrapeBill}}
}
