% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchBill.R
\name{searchBill}
\alias{searchBill}
\title{Search Bills}
\usage{
searchBill(key = NULL, congress = NULL, chamber = "Both",
  billtype = "bills", billstatus = NULL, subject = NULL,
  house_committee = NULL, senate_committee = NULL, party = NULL,
  pageSort = "documentNumber", pageSort_direction = "descending",
  maxN = 1000, urlonly = FALSE, searchURLprint = FALSE)
}
\arguments{
\item{key}{A \code{character} element of keywords.}

\item{congress}{A \code{character} vector of target congress(es) (e.g, \code{c(116,115)}).}

\item{chamber}{A \code{character} of target chamber. One of \code{"House"}, \code{"Senate"}, and \code{"Both"}. The default is both chambers.}

\item{billtype}{A \code{character} vector of target bill types. Choose any combination of 
\code{"bills"}, \code{"amendments"}, \code{"resolutions"}, \code{"joint-resolutions"}, and 
\code{"concurrent-resolutions"}. If \code{NULL}, include all types. The defauls is \code{"bills"}.}

\item{billstatus}{A \code{character} vector of target bill statuses. Choose any combination of 
\code{"introduced"}, \code{"committee"},\code{"floor"},\code{"failed-one"},\code{"passed-one"},
\code{"passed-both"},\code{"resolving"},\code{"president"},\code{"veto"}, and \code{"law"}.
If \code{NULL} (default), include bills of all statuses.}

\item{subject}{A \code{character} vector of policy subject areas. Choose any combination of subjects provided in 
\href{https://www.congress.gov/advanced-search/subject-policy-area}{Subject Area List}. If \code{NULL} (default), 
include all policy areas.}

\item{house_committee}{A \code{character} vector of House committees. Choose any combination of 
committees from \href{https://www.congress.gov/advanced-search/choose-committees?chamber=House}{House Committee List}.}

\item{senate_committee}{A \code{character} vector of Senate committees. Choose any combination of 
committees from \href{https://www.congress.gov/advanced-search/choose-committees?chamber=Senate}{Senate Committee List}.}

\item{party}{A \code{character} vector of parties of a sponsor. Choose any combination of 
\code{"Democratic"},\code{"Republican"},\code{"Independent"} and \code{"Independent Democrat"}.
If \code{NULL}, include any party.}

\item{pageSort}{The way the bills in search are sorted. Choose one of 
\code{"relevancy"}, \code{"title"}, \code{"documentNumber"} (default), 
\code{"lawNumber"}, \code{"latestAction"}, and \code{"dateOfIntroduction"}.}

\item{pageSort_direction}{Direction of \code{pageSort}. Choose \code{"descending"} (default) or \code{"ascending"}. 
Not applied if \code{pageSort} is one of \code{"relevancy"} and \code{"title"}.}

\item{maxN}{The max number of bills to import. The default is \code{1000}.}

\item{urlonly}{if \code{TRUE}, export character vector or bill summary page URLs. 
If \code{FALSE}, export \code{data.frame} of bill URLs and other extractable information from search page.}

\item{searchURLprint}{If \code{TRUE}, return search page URL.}
}
\value{
If \code{urlonly==FALSE} and \code{searchURLprint==FALSE}, 
A \code{data.frame} that contains following variables.
\itemize{
  \item \code{ID}: Bill ID (See \code{\link{BillIDfromURL}})
  \item \code{URL}: Bill summary page URLs
  \item \code{Title}: Bill title
  \item \code{Sp_Chamber}: Chamber (i.e., House or Senate) of sponsor
  \item \code{Sp_Name}: Sponsor full name
  \item \code{Sp_LastName}: Sponsor last name
  \item \code{Sp_Party}: Party of sponsor
  \item \code{Sp_State}: State of sponsor
  \item \code{Sp_District}: District of sponsor
  \item \code{Cosp_N}: Number of cosponsors
  \item \code{IntroducedDate}: Date of introduction
  \item \code{ActLatestDate}: Date of the latest action
  \item \code{ActLatestText}: Content of the latest action
  \item \code{Committees}: Name of committees
}
}
\description{
Searching Bill Information in \href{https://www.congress.gov}{US Congress Website}.
}
\examples{
# Search all bills (importing only newest 10)
billdt <- searchBill(maxN=10)
# Seach bills with keyword "air" (importing newest 100)
billdt <- searchBill(key="air", maxN=100)


}
\seealso{
\code{\link{getBill}}, \code{\link{read_scrapeBill}}, and \code{\link{write_scrapeBill}}
}
