% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scrapeBill.R
\name{read_scrapeBill}
\alias{read_scrapeBill}
\alias{write_scrapeBill}
\title{Read and write \code{scrapeBill} Object}
\usage{
read_scrapeBill(file)

write_scrapeBill(x, file, ...)
}
\arguments{
\item{file}{A connection or the name of the file (use \code{.rds} extension).}

\item{x}{A (or A list of) \code{scrapeBill} object created by \code{\link{scrapeBill}} function.}

\item{...}{Additional argument passed to \code{\link[base]{saveRDS}}}
}
\value{
For \code{read_scrapeBill}, an \code{R} object. 
For \code{write_scrapeBill}, \code{NULL} invisibly. \code{write_scrapeBill} also writes out 
auxiliary files under \code{FILENAME_files} folder. DON'T delete the folder. 
\code{read_scrapeBill} will need the files in the folder to read file.
}
\description{
Write (especially a list of) \code{scrapeBill} object.
}
\references{
\url{https://stackoverflow.com/questions/37703689/cannot-save-load-xml-document-generated-from-rvest-in-r}
}
